+++ mgmt-api/src/main/java/org/holodeckb2b/bdxr/smp/server/mgmtapi/xml/BusinessEntity.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2023 The Holodeck B2B Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Affero GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.holodeckb2b.bdxr.smp.server.mgmtapi.xml;
+
+import javax.xml.bind.annotation.XmlRootElement;
+
+import eu.peppol.schema.pd.businesscard._20161123.BusinessEntityType;
+
+@XmlRootElement(name = "BusinessEntity", namespace = "http://www.peppol.eu/schema/pd/businesscard/20161123/")
+public class BusinessEntity extends BusinessEntityType {
+
+}