+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### 
 ### Added
 * Support for Peppol Directory Business Cards
+* REST API to manage Participants
+* Option to select which servers (Query/AdminUI/REST API) are started
 * Specific exceptions for errors in communication with SML and directory
 
 ### Changed